move layer generation into layers
